requirements.txt
monai
torch --index-url https://download.pytorch.org/whl/cu128
torchvision
scikit-image
scikit-learn
nibabel
tqdm
PyYAML

